Switch from ws-rs to tungstenite-rs
