+++ Cargo.toml
@@ -21,7 +21,7 @@ cloudflare = "0.4.1"
 env_logger = "0.6.1"
 failure = "0.1.5"
 log = "0.4.6"
-openssl = "0.10.26"
+openssl = { version = "0.10.26", optional = true }
 reqwest = "0.9.18"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.39"
@@ -53,10 +53,11 @@ chrono = "0.4.9"
 tokio = "0.2.0"
 # chrome-devtools-rs = { path = "../chrome-devtools-rs" }
 chrome-devtools-rs = { git = "https://github.com/everlastingbugstopper/chrome-devtools-rs", rev = "b7a0e9f" }
+ws = "0.9.0"
 
-[dependencies.ws]
-version = "0.9.0"
-features = ["ssl"]
+[dependencies.tungstenite]
+version = "0.9.2"
+features = ["tls"]
 
 [dev-dependencies]
 assert_cmd = "0.11.1"
